blog modify and delete authorization

+++ start/routes/blog.py
@@ -41,10 +41,17 @@ async def get_blog_by_id(
     blog = await blog_svc.get_blog_by_id(conn, id)
     blog.content = util.newline_to_br(blog.content)
 
+    is_valid_auth = auth_svc.check_valid_auth(
+        session_user=session_user, blog_author_id=blog.author_id, blog_email=blog.email
+    )
     return templates.TemplateResponse(
         request=request,
         name="show_blog.html",
-        context={"blog": blog, "session_user": session_user},
+        context={
+            "blog": blog,
+            "session_user": session_user,
+            "is_valid_auth": is_valid_auth,
+        },
     )
 
 
@@ -93,6 +100,15 @@ async def update_blog_ui(
     session_user=Depends(auth_svc.get_session_user_prt),
 ):
     blog = await blog_svc.get_blog_by_id(conn, id=id)
+    is_valid_auth = auth_svc.check_valid_auth(
+        session_user=session_user, blog_author_id=blog.author_id, blog_email=blog.email
+    )
+
+    if not is_valid_auth:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="해당 서비스는 권한이 없습니다.",
+        )
 
     return templates.TemplateResponse(
         request=request,
@@ -111,6 +127,17 @@ async def update_blog(
     conn: Connection = Depends(context_get_conn),
     session_user=Depends(auth_svc.get_session_user_prt),
 ):
+    blog = await blog_svc.get_blog_by_id(conn, id=id)
+
+    is_valid_auth = auth_svc.check_valid_auth(
+        session_user=session_user, blog_author_id=blog.author_id, blog_email=blog.email
+    )
+
+    if not is_valid_auth:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="해당 서비스는 권한이 없습니다.",
+        )
     image_loc = None
     author = session_user["name"]
     if len(imagefile.filename.strip()) > 0:
@@ -136,9 +163,21 @@ async def update_blog(
 
 @router.delete("/delete/{id}")
 async def delete_blog(
-    request: Request, id: int, conn: Connection = Depends(context_get_conn)
+    request: Request,
+    id: int,
+    conn: Connection = Depends(context_get_conn),
+    session_user=Depends(auth_svc.get_session_user_prt),
 ):
     blog = await blog_svc.get_blog_by_id(conn=conn, id=id)
+    is_valid_auth = auth_svc.check_valid_auth(
+        session_user=session_user, blog_author_id=blog.author_id, blog_email=blog.email
+    )
+
+    if not is_valid_auth:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="해당 서비스는 권한이 없습니다.",
+        )
     await blog_svc.delete_blog(conn=conn, id=id, image_loc=blog.image_loc)
     return JSONResponse(
         content="메시지가 삭제되었습니다", status_code=status.HTTP_200_OK
